Added DotSDL audio support.
    * OpenAL calls replaced with SDL callbacks.
    * Roton changed to internally use floats for audio samples.

+++ Roton.Interface/Audio/AudioComposer.cs
@@ -49,7 +49,7 @@ namespace Roton.Interface.Audio
             _bufferNumerator = outputSampleRate * 10;
         }
 
-        private IEnumerable<int> ComposeAudio()
+        private IEnumerable<float> ComposeAudio()
         {
             while (_bufferAccumulator > _bufferDenominator)
             {
